server/requirements.txt
# requirements.txt
Adafruit_GPIO
Adafruit_MCP3008
rpi-gpio
spidev
gpiozero==1.5.1
Flask==1.1.1
openpyxl==3.0.3
rethinkdb==2.4.5
